backend/requirements.txt
asgiref==3.11.1
Django==4.2.30
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
PyJWT==2.12.1
sqlparse==0.5.5
typing_extensions==4.15.0
